requirements.txt
openai>=1.0.0
python-dotenv==1.2.2
streamlit>=1.32.0
chromadb>=0.5.0
sentence-transformers>=2.6.0
pdfplumber>=0.10.0
python-docx>=1.1.0
